Close a file output stream when done with it
The garbage collector and operating system would clean this up for us
eventually, but it's good practice to close file streams promptly as
soon as we are done writing to them.

Unfortunately, Gradle doesn't give us a good way to finalize a task,
i.e., to run some cleanup code after the task ends even if it fails.
`Task.finalizedBy` requires a whole other task, and there's no easy way
to pass the open `FileOutputStream` from the execution task to some
other finalizer task.  So we're going to settle for best-effort here:
the file output stream is closed promptly when the subprocess succeeds,
but may linger until the next garbage collector run if the task fails.
That's at least better than what we were doing before.