Record Git commit ID in published build scans
I occasionally use `./gradlew --scan ...` to get deeper insight into the
effects of my changes to our build logic.  Often that means comparing
scan resolve before and after some candidate change.  However, it's very
easy to lose track of which scan corresponded to which version of a
change that I'm working on.

This change adds "Git Commit ID" as a custom value in each published
build scan.  You can see this custom value on the "Custom values" tab
when viewing a build scan in your web browser.  Its value is the current
Git commit hash, with `-dirty` appended if there were uncommitted
changes in the tree at the time of the build.