Don't check Eclipse settings for typos
These files contain a few real spelling errors that we would want to
correct if they appeared _anywhere_ else, such as "statment" and
"delcaration".  My guess is that these were uncaught typos in the
original Eclipse implementation, and now _everyone_ is stuck with them
because they've become the names of specific configuration parameters.