Simplify a URL utility function
Personally, I'm skeptical of the nonstandard normalizations that
`relativeToAbsoluteUrl` initially applies to `urlFound`:  replacing
`\` with `/` and converting to lower case.  But that's what was done
before, presumably for good reason.

Note that the new implementation of `relativeToAbsoluteUrl` differs
from the old implementation in how it handles `..` in the interior of
the `urlFound` path.  The new implementation normalizes those away,
whereas the old implementation did not.  I'm OK with this change, so
I've updated the test suite accordingly.