Add some "see also" links to a comment
The first link is to a Gradle discussion forum exploring alternative
approaches.  I think that the best solutions offered here are
educational but overkill for our current needs.

The second link is to a Gradle feature request that would make a robust
implementation much easier.  If that feature is added in the future,
then we can adopt it here.