Update for JDK 25 (#1673)
As usual, we just test that previous Java bytecode support works, but
can't promise support for new features.

We needed a couple of code changes on this one. First, apparently `new
File("").exists()` returns true on JDK 25 (??). Not sure why we were
doing that before anyway; we fix up the Dalvik test code to just pass
`null` (and annotate a couple of methods to indicate that's legal).

Second, the `TestListIterator` slicer test stopped running in reasonable
time, presumably due to implementation-level changes in the JDK. We
rewrite the test to be self-contained and hence robust to these internal
JDK changes.

Finally, due to some Gradle toolchain issues, we had to tweak the CI
config to always install JDK 25, so we can use it to compile our code.