Improve links to standard Oracle Javadoc (#1706)
Previously, Javadoc for the `util` project linked to Oracle's
documentation for standard library components. That's a very useful
feature, not just for `util`, but for all of our documentation. Now we
enable this feature everywhere.

Previously we linked to Oracle's documentation for the version of Java
that was running Gradle itself. Now we use the version of Java that
corresponds to the current Java toolchain, which includes the Javadoc
tool itself.

Due to decisions made quite a while ago, that Java toolchain will always
be Java 11 or later. So there's no longer any need to generate links
into Oracle's Javadoc for Java 10 or earlier.