Factor out helpers for using version catalogs (#1707)
Gradle generates type-safe accessors for version catalog entries.
Unfortunately, those accessors are not available to precompiled script
plugins, which requires the latter to use lower-level APIs. We use those
lower-level APIs quite a bit in our `build-logic` code. They are not as
friendly as the generated accessors, but we can do a few things to make
using them more convenient.

While we're doing that, we can also provide friendlier error messages if
a requested catalog entry cannot be found. That never seemed worth doing
at any individual place where we were using the lower-level APIs. But
now that we're factoring this code out and implementing it just once,
it's worth the additional effort to make that code as helpful as
possible.