Fix project vs. publication group confusion (#1709)
A `MavenPublication` has a `groupId`, which defaults to the current
`Project`'s `group`. A `MavenPublication` does not have a `group`,
though. So when we were assigning to `group` inside a `publishing`
block, we were actually changing the `group` of the `Project`, which is
not a sensible or useful thing to do here.