Defend test against vacuous success and order changes
An earlier version of this test passed vacuously because it was looking
for a node named `"call4"` instead of `"$$ call_4"`.  That mistake has
already been corrected.  However, we can make this test even more robust
by ensuring that the sought-after node appears exactly once.

Furthermore, when validating the contents of a list of successors, we
should not assume that those successors are given in any particular
order.

See <https://github.com/wala/WALA/pull/1726> for further discussion on
recent corrections and improvements to this test.