Stop using a deprecated `Runtime.exec` overload
The deprecated overload of this method takes a single `String` as the
command to execute, which it parses into a command vector using a
`StringTokenizer`.  The preferred (non-deprecated) overload takes a
command vector directly as a `String[]`.  The latter is less
error-prone.  Conveniently, we _do_ already have a way to get at the
correct `String[]` command vector.