Leverage Guava's bidirectional map type
The `CAstToNode` and `nodeToCAst` fields of a `CAstControlFlowRecorder`
instance were already both one-to-one maps and inverses of each other.
Guava's `BiMap` is perfectly suited to managing that kind of data.
`BiMap` also makes explicit the relationship between these two maps,
which was otherwise just implied by their names and types.

The change to `smokeMainExtraPathElements` in
`cast/smoke_main/build.gradle.kts` is needed now to clarify that we want
JRE variants of dependencies.  Guava is available for JRE and for
Android.  In theory, either should work in either environment, though
each might work better in the environment for which it is optimized.
Without this clarifying attribute, though, Gradle cannot pick between
the two and therefore dependency resolution fails.  With this attribute,
Gradle knows that we prefer the JRE variant, not the Android variant.