Create fewer `Factory` and `Function` instances
Each call to `findOrCreateValue` needs a `Factory` instance.  The body
of `findOrCreateValue` adapts that factory into a `Function` instance
that implicitly captures the factory from the enclosing lexical
environment.  That adds measurable overhead, especially if the common
case is that the key is already present.  The other `findOrCreate*`
methods can reduce that overhead by skipping `findOrCreateValue` and
instead calling `Map.computeIfAbsent` directly.