Allow an `assert` with a side effect (#1805)
IntelliJ IDEA warns that this `assert` has side effects. In general,
that's a likely bug because these effects would happen or not depending
on the JVM's global `assert`-checking mode. However, in this _specific_
case, the possible side effect is safe, because `succs` is never
subsequently used after the `assert`.