Avoid redundant map lookups in `cast/js/rhino` (#1813)
Remove some redundant `Map` lookups, such as calling `containsKey`
immediately before `get`. That's redundant as long as we never map to
`null`, which appears to be the case for the maps we're modifying here.

Also codify the existing `null`-avoidance policy for these maps by
adding JSpecify annotations. It's not clear how thoroughly current tools
check this annotation when used on a `Map`'s value type. But even if
current tools don't check them, it's worth adding these annotations as
documentation and (hopefully) as hints for future checkers.