Avoid redundant map lookups in `core` (#1821)
Remove some redundant `Map` lookups, such as calling `containsKey`
immediately before `get`. That's redundant as long as we never map to
`null`, which appears to be the case for the maps we're modifying here.

Also codify the existing `null`-avoidance policy for these maps by
adding JSpecify annotations. It's not clear how thoroughly current tools
check this annotation when used on a `Map`'s value type. But even if
current tools don't check them, it's worth adding these annotations as
documentation and (hopefully) as hints for future checkers.

Conversely, add `@Nullable` to the value type for some maps that _can_
have `null` values. This would be the assumed default anyway. However,
deciding whether `null` values are possible or not takes some effort. If
I've invested that effort and determined that they are possible, then it
is useful to write that down explicitly. Otherwise, someone else might
have to go through the same effort again in the future.